visustat_check

Gives a summary that allows to check the linking between a passed dataframe and images in a frame-wise manner.

visustat_check(df, images, stack_time = FALSE, dimensions = 2)

Arguments

df
dataframe of the form: df(track, time, X, Y, (Z,) mapping_parameters, ...)

images
vector: filenames of images ordered by frames/time

dimensions
numeric: specify whether the images are 2D or 3D. If 3D is selected data is assumed to be in the form: df(track, time, X, Y, Z, mapping paramters, ...)

stack
logical: default: FALSE, single image file provided if time-resolved imagestack is used, set: TRUE

Examples

data(hiv_motility)
images <- hiv_motility_images()
visustat_check(hiv_motility, images)
#> Same number of images as timepoints in the dataset.
#> # A tibble: 15 × 3
#>     time frame image                                                            
#>    <dbl> <int> <chr>                                                            
#>  1     2     3 /private/var/folders/cf/074kcbdn6_5879bsv0hg2qh00000gn/T/Rtmpja1…
#>  2     3     4 /private/var/folders/cf/074kcbdn6_5879bsv0hg2qh00000gn/T/Rtmpja1…
#>  3     4     5 /private/var/folders/cf/074kcbdn6_5879bsv0hg2qh00000gn/T/Rtmpja1…
#>  4     5     6 /private/var/folders/cf/074kcbdn6_5879bsv0hg2qh00000gn/T/Rtmpja1…
#>  5     6     7 /private/var/folders/cf/074kcbdn6_5879bsv0hg2qh00000gn/T/Rtmpja1…
#>  6     7     8 /private/var/folders/cf/074kcbdn6_5879bsv0hg2qh00000gn/T/Rtmpja1…
#>  7     8     9 /private/var/folders/cf/074kcbdn6_5879bsv0hg2qh00000gn/T/Rtmpja1…
#>  8     9    10 /private/var/folders/cf/074kcbdn6_5879bsv0hg2qh00000gn/T/Rtmpja1…
#>  9    10    11 /private/var/folders/cf/074kcbdn6_5879bsv0hg2qh00000gn/T/Rtmpja1…
#> 10    11    12 /private/var/folders/cf/074kcbdn6_5879bsv0hg2qh00000gn/T/Rtmpja1…
#> 11    12    13 /private/var/folders/cf/074kcbdn6_5879bsv0hg2qh00000gn/T/Rtmpja1…
#> 12    13    14 /private/var/folders/cf/074kcbdn6_5879bsv0hg2qh00000gn/T/Rtmpja1…
#> 13    14    15 /private/var/folders/cf/074kcbdn6_5879bsv0hg2qh00000gn/T/Rtmpja1…
#> 14     0     1 /private/var/folders/cf/074kcbdn6_5879bsv0hg2qh00000gn/T/Rtmpja1…
#> 15     1     2 /private/var/folders/cf/074kcbdn6_5879bsv0hg2qh00000gn/T/Rtmpja1…